ml/requirements.txt
# extraction.py
requests
python-dotenv
langchain
dataclasses
typing

# embed.py
# sentence_transformers
# langchain-community
# nltk
# torch

#emotion.py
mediapipe